docker/requirements.txt
Flask==2.0.2
requests==2.26.0
Pillow==8.3.1
numpy
scikit-learn